[refactor] Controller 역할을 수행하는 Game 클래스의 패키지 이동
